+++ apps/api/src/secret/secret.service.ts
@@ -1,5 +1,6 @@
 import {
   ConflictException,
+  forwardRef,
   Inject,
   Injectable,
   Logger,
@@ -41,6 +42,7 @@ import { SecretWithProject, SecretWithValues } from './secret.types'
 import { AuthenticatedUser } from '@/user/user.types'
 import { TierLimitService } from '@/common/tier-limit.service'
 import SlugGenerator from '@/common/slug-generator.service'
+import { VariableService } from '@/variable/variable.service'
 import { encrypt } from '@keyshade/common'
 
 @Injectable()
@@ -53,6 +55,8 @@ export class SecretService {
     private readonly authorizationService: AuthorizationService,
     private readonly tierLimitService: TierLimitService,
     private readonly slugGenerator: SlugGenerator,
+    @Inject(forwardRef(() => VariableService))
+    private readonly variableService: VariableService,
     @Inject(REDIS_CLIENT)
     readonly redisClient: {
       publisher: RedisClientType
@@ -94,6 +98,9 @@ export class SecretService {
 
     // Check if the secret with the same name already exists in the project
     await this.secretExists(dto.name, project)
+
+    // Check if a variable with the same name already exists in the project
+    await this.variableService.variableExists(dto.name, project)
 
     const shouldCreateRevisions = dto.entries && dto.entries.length > 0
     this.logger.log(
@@ -233,6 +240,10 @@ export class SecretService {
 
     // Check if the secret with the same name already exists in the project
     dto.name && (await this.secretExists(dto.name, secret.project))
+
+    // Check if a variable with the same name already exists in the project
+    dto.name &&
+      (await this.variableService.variableExists(dto.name, secret.project))
 
     // Check if the user has access to the environments
     const environmentSlugToIdMap = shouldCreateRevisions
@@ -1156,7 +1167,7 @@ export class SecretService {
    * @param secretName the name of the secret to check
    * @param project the project to check the secret in
    */
-  private async secretExists(secretName: Secret['name'], project: Project) {
+  async secretExists(secretName: Secret['name'], project: Project) {
     this.logger.log(
       `Checking if secret ${secretName} exists in project ${project.slug}`
     )
@@ -1175,5 +1186,8 @@ export class SecretService {
         constructErrorBody('Secret already exists', errorMessage)
       )
     }
+    this.logger.log(
+      `Secret ${secretName} does not exist in project ${project.slug}`
+    )
   }
 }